refactor: move log_importer under sled_compat
# Summary

The sled-to-WAL importer now lives in the module that holds the rest of
the sled compatibility code.

# Details

`log_importer` reads `RaftStoreEntry` and `RaftStateKey` from
`sled_compat` and has a single caller, the v003-to-v004 upgrade. Sitting
at the crate root it read like a general facility rather than the
migration helper it is.

Moving it also leaves `db_exporter` alone at the root, which is an
accurate picture: it is the only live, non-compatibility code in
`store-compat`, and the export format has to be versioned before it can
move anywhere else.
